<header> en <footer> refractored en vanuit main in app.jsx geplaatst. Uit de header de nav gehaald: ik laat deze in Header.jsx staan om de layout correct te houden
