+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,11 @@
+
+.buttons-container {
+    display: flex;
+}
+
+
+.disabled {
+    pointer-events: none;
+    opacity: 0.6;
+    cursor: default;
+}
